slurmd - properly handle whole het job when one component fails to launch
When a launch failure occurred on one het component job, most notably
when the node prolog failed but also potentially under other conditions,
only the component was handled and not the whole het job. This patch fixes
that problem by operating on the whole het job when one comp job fails.

Bug 13747