Alter slurm_cred_faker() to use common code paths.
The only thing "fake" about these credentials is the signature block.
Since cred/munge will work for any user on the system, we can use a
"real" signature - it will just be signed by the calling user instead.

(This "fake" mechanism was required for cred/openssl, when the client
commands would not have been able to generate valid signatures.)