Move common code into setup_cred_arg().
create_kill_job_msg() does not do much setup as it was just added, but
start populating the fields anyways in anticipation of a longer-term
pivot over to this.

Also, successive commits will introduce some additional fields, and it's
much simpler to keep placing them centrally in the job credential than to
modify a wide array of other RPCs.