Create job node_inx's cluser_nodes instead of node_ptr->index
node_record_table_ptr can have NULL slots and result in bitmaps that
don't match the hostlist that the dbd needs.
.e.g.
node_record_table_ptr
[0]=node1
[1]=NULL
[2]=node2
job runs on node1 and node2.

The bitmap generated against node_record_table_ptr would be 0,2. But the
dbd doesn't know about the NULL slots and expects node[1-2] to be 0,1. A
query for jobs running on node2 won't be found because node2 on the
controller is index 2 and on the dbd it's index 1.

So generate node indexes based off of a hostlist that doesn't know about
the holes in node_record_table_ptr.

Bug 13679