Testsuite - Adapt test31.3 to 76409450c5 replacing SLURM_JOB_CONSTRAINTS
Prior behavior of this test was to check if the
SLURM_JOB_CONSTRAINTS env var was in the prolog.out even
though it showed up as SLURM_JOB_CONSTRAINTS=null.

Because some env vars showed when null and others didn't, 76409450c5
made the decision across the board not to show if null, thus breaking
the test when it no longer appeared.

This patch replaces that check with one for SLURM_JOB_COMMENT instead
to avoid potential negative impacts of having to set constraints /
features for the test.

Bug 13958