task/cgroup memory now works at task level
There was a race when jobacctgather/cgroup moved a pid to task_x directory
in the memory controller. Then task/cgroup was removing it and putting again
the pid in the step_0 superior directory.

Do here what we already do for task/cgroup devices, and work at task level.

The conflicting controllers between task/cgroup and jobacctgather/cgroup are
only memory, since the cpu is gathered from cpuacct and task works with
cpuset.

Bug 13697