Fix improperly using return codes for run_command_waitpid_timeout()
This function was documented:

 *  Same as waitpid(2) but kill process group for pid after timeout secs.

However, the function returned 0 on success, not the pid of the process.
Also, callers of this function expected a non-zero exit status to
indicate a timeout. But this is not how the function worked, nor is it
how waitpid(2) is documented to work.

So, this commit:
* Fixes how return codes are set and used for this function.
* Adds an output parameter to this function to determine if it timed
  out while waiting for the script to terminate.

Bug 13967