Remove xassert
With changes in Bug 12146, a job can be requeued before
_accounting_cluster_ready() is called, thus cluster_hl won't be set yet.

job_requeue()
    ...
    acct_policy_update_pending_job()
        ...
        acct_storage_p_node_inx()

In this case where the job is being requeued, the job still has "nodes",
loaded from state, set and won't be cleared until the job is killed on
any remaining nodes and batch_requeue_fini() clears "nodes".

If cluster_hl is NULL, then the acct_storage_p_node_inx() will just
return NULL because the "nodes" string won't be found in cluster_hl.
This will be fine because cluster_hl should always be set before any
jobs are started. In this case, where the job is requeued before
cluster_hl is set, the job is pending from the requeue and is just
creating a new record in the db for the requeued job. A new start
message will be sent when the job is actually started.

Continuation of f97091000da