Avoid setting r_uid to SlurmdUser/SlurmUser on replies.
This was already in the 20.11 version of the security fix, but I neglected
to include it with the 21.08 version.

Work around a problem with a mix of patch + unpatched slurmd processes
when using message forwarding.

There is no point to setting r_uid to SlurmdUser/SlurmUser anyways, as
if we leaked these auth tokens to an untrusted user these would be
usable to impersonate SlurmdUser to slurmd or SlurmUser to slurmctld.