Set r_uid to sender msg->auth_uid in _resp_msg_setup().
Don't set the restrict uid if the destination is SlurmUser/root, as this
will cause communication problems to unpatched older slurmd processes.
Longer-term this code should be removed anyways, as there's no point in
sending the auth token on an already-established communication tree.

This is the issue that was fixed by 21.08.8-2 with cc962db09b.

Co-authored-by: Dominik Bartkiewicz <bart@schedmd.com>