Retroactively change the RPC for stepd_attach().
Unfortunately we needed to add this in as part of the security fix so
an appropriate r_uid value can be used, but this did technically
break the older RPC formats. Fortunately this RPC is only used for
sattach, so the potential impact is low and was deemed less severe
than leaving this code path un-fixed.