Change squeue --states=all to include all states
Modified the behavior of the --states=all option in the squeue command
to not depend on delineating each state. An all_states parameter now
exists as a flag to mark if we are printing all states or not,
making it robust to the adding or deleting of slurm states in other
parts in the code.

Bug 13678