Fix corner case bad assoc_ptr reference.
Reproducer:
start a job that will hang out after cancel (I just hacked slurmd to
ignore signals to the step)
scancel the job - it will keep running
Remove it's association. The DBD will allow this to happen because the
job end already made it to close the job out as CANCELED.
Then issue a scontrol reconfig.

You will notice the check of IS_JOB_FINISHED() is there to not double
call acct_policy_remove_job_submit() as it has already happened.  But
it doesn't clear out the association pointer which needs to happen.