Fixed logging when node goes from valid to invalid
Fixed the error when a user manually sets node to DOWN or DRAIN
with a custom "reason" and sees nothing when a node error occurs and
puts it in the invalid state. Now the error will appear in the logs
at least. As of now, we are not touching the custom reason the user set.

Bug 13668