Fix slurmctld segfault when a step requests a gres without a file
Reproducer:
* Allocate a job with a gres that does not have a file
* Run a step using that gres

gres_bit_alloc is only allocated for gres that have files. We were
dereferencing gres_bit_alloc when it was NULL (for a gres without a file),
causing the segfault.

gres_cnt_node_alloc can be used to know if the step was allocated any
gres on this node.

Bug 14079

Co-authored-by: Scott Hilton <scott@schedmd.com>