Use slurmd's hostname for dynamic future nodes
In 22.05, slurmd passes it's hostname in the registration message. Use
this instead of assuming the node name is the hostname because it can be
different.

e.g.
n1$ slurmd -F -Nblah1
blah1 doesn't match any nodes in the slurm.conf but matches an available
future node. Previously, the NodeHostname would be set to blah1 instead
of the actual hostname n1.

Bug 13679