Testsuite - Update test because gres are now in task cgroup
Because CUDA_VISIBLE_DEVICES is always 0, because the task is in a
cgroup, test1.119 should be run without Constrain devices so we can know
what gpus are being assigned to each task.
I believe that we wanted constrain devices before because that would
mean the gpus would be just 0,1. But now other numbers could be used if the
node has other jobs running.

Bug 14111