Fix race caused by f18b8c2f91 in test1.23.
This also simplifies the print here as %ps handles step_id = NO_VAL.

We only want to print the Force Terminated if we really were.  Before
we could race into this we were just denied an allocation for one reason
or the other.

Continuation of f18b8c2f91

Bug 14108