Don't limit the size of TaskProlog output
TaskProlog's output was previously limited to 4094 characters per line. This
limited the size of environment variables that could be exported to the task
via TaskProlog, or the size of a logging statement printed to the task via
TaskProlog.

Use run_command() which already handles this correctly (it resizes the
buffer storing the script's output as needed).

Bug 14187