Don't show message "Controller %s is not enabled!" when started from cmd
When slurmd is started without systemd some controllers might not be
present in the cgroup that slurmd is spawned in. An error was shown,
but it was not a real error as later slurmd will be moved to the
slurmstepd cgroup which has the proper controller available. This commit
converts that error into an info message if not started by systemd.

Bug 14246