Remove filtering for sprio from slurmctld
Filtering jobs for sprio is now done on the client-side rather than on
the controller, so this commit removes any filtering on the
controller. This commit also ensures that the controller will remain
backwards compatible with previous versions of sprio that still expect
the controller to filter jobs. The backwards compatibility code
should be removed in 24.08

Bug 14210