Allow powersave scripts to run for 10 seconds before shutting down
This preserves the behavior of running powersave scripts in slurmctld
(without slurmscriptd).

Prior to running powersave scripts in slurmscriptd, we waited up to 10
seconds for powersave scripts to complete before shutting down.
When calling ResumeProgram, a temporary file is opened in memory, and
that file will be closed once the daemon shuts down. Wait to allow
ResumeProgram scripts to read the contents of that file while it exists.

Bug 12058