Fix logic for distributing tasks among heterogeneous nodes in cons_tres
Previously, if there was a heterogeneous set of nodes in a partition with
differing cpu counts and you submitted a job that would exahust resources
on one node before the others, the job would not be scheduled because of
inadequate resources, even though between all the nodes it could be
scheduled. This only happened when using the -N option, otherwise the job
would fill all cpus between all nodes just fine.
This patch removes the rounding up division that was causing this issue.

Bug 14262