openapi/dbv0.0.38 - default preempt_mode to off
slurmdb_init_qos_rec() will set preempt_mode to NO_VAL which results in
no change in modify request. This default setting causes issues as the
flags attempt to get ORed into it pointlessly. With slurmrestd, it is
expected that the full definition will be provided for the QOS object
and not just requested changes.

Bug 14388