Docs - Add information about loginctl linger
This section adds information on how to have the features of the
pam_systemd module without using it. Executing loginctl enable/disable
linger in the prolog and epilog all the user management and environment
preparation gets done in a very similar way as if the pam_systemd module
was still there. Added also prolog, epilog and task_prolog examples.

Bug 14392