openapi/dbv0.0.38 - fix QOS flag modifications
If QOS_FLAG_NOTSET is set, then the mysql plugin skips the code which
overwrite flags. By not setting this flag, we ensure that the mysql plugin
will always overwrite QOS flags with whatever is specified.

Continuation of the previous commit.

It was found that this was a much simpler way to accomplish what commit
38ea9484c8 was originally doing, so the previous commit reverted
38ea9484c8 and this commit fixes it in a different way.

Bug 14388

Signed-off-by: Nathan Rini <nate@schedmd.com>