Docs - State that SPANK plugins are needed on login nodes
Adds a note stating that the SPANK plugins need to be installed on the
compute nodes as well as on the login nodes (or wherever you want to
run sbatch, srun and other job allocation commands). Also modifies the
message that says that slurm will terminate the job if something
strange is found in the plugin to include a reference to the job
allocation commands there.

Bug 14483