Fix expected ordering for hostlists from controller and daemons
The clients/daemons needs to be able to find hosts in the same order
that the controller is working off of. If the controller's
node_record_table_ptr order and a sorted hostlist's order differ, then
the client/daemon will incorrectly index into the wrong node's
information.

See also 11db9adbef.

Bug 14013