Fix gres_per_step based node allocation in heterogenous jobs.
When running gres_step_test() here we aren't testing a specific allocation's
group of nodes. But instead, independently testing each node to know if it
could be in the step.
Decrementing max_rem_nodes in this location means we don't test all the nodes
properly. It should be the max for each test so each test is independent.

To properly follow gres_per_step a separate check evaluating all nodes together
is required.

Bug 12996