cons_tres - Don't attempt socket removal for binding of any sock GRES
If the GRES is available on any socket we have to treat is as a per node
resource. There is no reason to remove cores from so called non-required
sockets, since all sockets are equally usable by the GRES.

The modified if statement only applied to a condition with
--gres-per-node specification, however, it applies to --gpus request
(reflected by max_gres, see _set_max_node_gres()) too.

Bug 14175