Make all cgroup directories created by Slurm owned by root
This was the behavior in cgroup/v2 but not in cgroup/v1 where by default
the step directories ownership were set to the user and group of the
job.

This prevents users/jobs from manipulating the Slurm cgroup hierarchy.

This can be reversed by using the undocumented RootOwnedCgroups
cgroup.conf paramter.

Bug 13964