openapi/dbv0.0.38 - Fix qos list/preempt not being parsed correctly
Changes the _add_func call related to PARSE_QOS_PREEMPT_LIST so that the
parsing now is done with _parse_qos_id_list implemented in commit
f0b23d1c77. In consequence removes _parse_qos_preempt_list as is not needed
anymore.

Also removes the special add_parse function add_parser_qos_preempt, now a
default _add_parse one is used, to do this the _dump function needs to not
assert on a non-zero parse->field_offset.

The function _update_qos retrieves the qos list from the DB and adds it
to the foreach_update_qos_t args, so that it can be later added to the
parser_env_t that is sent to the parse function _parse_qos_id_list.

Bug 14498