Make logs more helpful
A misleading error message is returned when a user is not allowed to
reboot nodes with the node features plugin and requests a valid
changeable feature that is not active:

"Invalid feature specification"

This is misleading because the feature is valid but not active, so the
user cannot use this feature.

This commit modifies the slurmctld log message to make it easier to
diagnose the problem.

Bug 14478