conmgr - avoid lazy linking of host parser functions
The lazy linking previously used for parse_host_port() and
free_parse_host_port() would result in errors when configuring with
'--without-shared-libslurm'.

To two functions are now passed as pointers to init_con_mgr() to force
the caller to explicitly provide the pointers at startup. This allows
complete removal of the stubs previously placed in http.c.

These remain as callbacks since, for slurmrestd, they are provided
by libhttpparser, and we do not want to force that library dependency
on libslurm (and thus globally to Slurm) at this time.

Bug 14595