cgroup/v1 - detect OOM events via memory.oom_control oom_kill when exposed
Previously, cgroup/v1 used a single mechanism to detect OOM events. This
was done by registering a notifier via creating an eventfd, opening the
memory.oom_control file and writing a string like "<event_fd> <fd of
memory.oom_control>" to cgroup.event_control. Then, a new thread was
created to poll and read a counter of OOM events from the eventfd.

With this mechanism, if a step requested the same amount of memory as
the job allocation, and any step task memory usage exceeded the limit,
the OOM kill event was properly catched and accounted for the step, but
sibling steps would also incorrectly be notified and accounted as OOM even
if they had nothing to do with it.

Kernel versions > 3.x expose an oom_kill field that can directly be read
from the memory.oom_control interface. This patch makes it so that when
this is available, use this mechanism instead of the previous eventfd +
thread creation one, fixing the incorrect accounting issue for the
siblings in that specific use-case.

It's also needed to account for each task in the step at step termination
because the counter is not hierarchical unlike in cgroup/v2.

Bug 9737