Fix dynamic future node registrations
Multiple registrations registrations were getting mapped to the same
node name. This was because the assigned node was still being treated as
available to be mapped because it's state was still FUTURE.

After assigning the slurmd to a Slurm node name, change the
assigned node's state to FUTURE so it won't get found again.

Bug 14586