Allow shared gres to use higher gres counts.
We have an error check for this both here and in _node_config_validate.
This commit makes them match.

Shared gres bitmaps are the size as the gres they are splitting up not the size
of their count.
We check MAX_GRES_BITMAP when initalizing the host gres.
We therefore shouldn't limit the count of shared gres.

Bug 14635