Do not limit avail_cpus before we have picked cores
Make avail_cpus actually be the number of CPUs available, not the number
of CPUs required for the job.

This avoids picking cores too early before we know how many sockets and
which sockets we need, in particular for GRES on different sockets.
Later logic actually picks cores for the job.

For example, without this commit, a job requesting GPUs on different
sockets will still be given cores on one socket. If that job also
requested --gres-flags=enforce-binding and gres.conf has Cores set
appropriately for each GPU, the job is wrongly rejected. This commit
means that job is no longer rejected.

Bug 14153