Do not kill script if there is no more data to read from its stdout
If a script closes stdout (or redirects stdout in a way that closes it),
this does not mean the script is done. Before this commit, run_command()
used the status of the file descriptor to determine if the script had
finished. With this commit, run_command() will stop calling poll() when
there is no more data to read and will wait for the script to finish,
timeout or be killed by track_script, or for shutdown.

Bug 14457

Signed-off-by: Felip Moll <felip.moll@schedmd.com>