Gres types without plugin work with the function g_get_devices
Create a new struct with the gres name and a list of the gres_devices
associated with that gres. Then a list of that structs is created, this
list holds all the devices that are not backed by a plugin. This list is
created is slurmd at gres_g_node_config_load, packed and later sent to
slurmstepd, so that when stepd receives a g_get_devices, for the devices
that do not have a plugin, it can get those from this list.

Bug 14139

# Conflicts:
#	src/common/gres.c