switch/hpe_slingshot - use new cxil_get_svc_rsrc_list routines
PE-42187 - use new cxil_get_svc_rsrc_list routines in plugin

When creating a CXI service for an application on a compute node,
the Slurm Slingshot switch plugin looks at any system services
already on the node, and subtracts the reserved value for each
limit from the max available on the node.  It then uses that
adjusted limit as a max value for any resource request from
an application launch.  This allows applications with large thread
counts to proceed instead of failing during service creation.

It turned out that certain system services were using more
resources than their reserved limits (legal as long as the in_use
is less than the max limit).  A new libcxi function was added
to get the in_use value for a particular service.

This changes uses the new libcxi function, and will subtract
MAX(reserved limit, in_use) from the maximum for each resource
on the node. Workaround older version of libcxi that do not have
this functionality.

Also in this mod, remove the "LIBCXI_1.0" version suffix from
the plugin; it doesn't work with libcxi versioning.

Tested on sawmill.

Bug 14604