Sort job and step nodelists in Slurm user commands
slurmctld does not send sorted nodelists anymore, so this commit preserves
the behavior of Slurm user commands showing sorted nodelists for jobs and
steps. This commit does it in the user commands rather than slurmctld for
performance reasons.

This means that 22.05 user commands without this commit, or user commands
prior to 22.05, that are talking to an updated 22.05 (or newer) slurmctld
will show unsorted nodelists for jobs and steps.

Bug 14013