Do not update node's available features if there was no change
orig_features already had the features from the plugin, so it was incorrect
to append the plugin features at the end of orig_features. This commit
corrects orig_features so that the original features can be compared to the
new features.

Avoiding updating node features when it is unnecessary also avoids splitting
a new config record.

Bug 14788

Signed-off-by: Brian Christiansen <brian@schedmd.com>