Prevent jobs from being scheduled on future nodes
A node can be put into the future state after slurmctld sends a
ping/registration message and before slurmctld receives the response.
When the slurmctld receives a response to the ping/registration message,
it would put the node into avail_node_bitmap and up_node_bitmap, allowing
jobs to be scheduled on the node. But those jobs would fail because the
node is in the FUTURE state - slurmctld does not know how to communicate
with those nodes, and jobs shouldn't run on FUTURE nodes anyway.

This commit prevents FUTURE nodes from being added to avail_node_bitmap
and up_node_bitmap

Bug 14830