Do not count specialized CPUs in a node's alloc CPUs
When a job requests specialized CPUs (with --core-spec or --thread-spec),
a whole node is allocated even though the job cannot use the specialized
CPUs.

When a whole node is allocated, all the CPUs on the node are used,
including the specialized CPUs. However, specialized CPUs are not
counted in the number of configured CPUs on the node, so do not count
them in the number of allocated CPUs on the node.

Bug 14319

Signed-off-by: Danny Auble <da@schedmd.com>