Don't pthread_join() a thread that was not started
In commit 278546d411f we do not start the power save thread if power save is
not enabled in the configuration. Avoid calling pthread_join() if the power
save thread was not started.

Calling pthread_join() with a thread id that does not exist segfaults in
CentOS/RHEL instead of returning ESRCH, so this also avoids a segfault in
CentOS/RHEL when slurmctld shuts down.

Bug 13953