Remove throttle code for accounting update
Now that we are only adding to a list and broadcasting a signal
we should be able come in here as many times as we like as everything
should be kept in order.

The persistent connection will keep this from being called multiple
times in parallel in the first place.

Bug 14864