slurmctld - Fix case where het jobs could over allocate licenses
A het job could cause license overallocation by distributing license requests
over the component jobs. Example: Given that there are 2 lic1 licenses, this
job would not fail to be submitted and 3 lic1 licenses would be assigned
(1 on the first component job and 2 on the second):

sbatch -L lic1:1 : -L lic1:2 ...

This patch fixes the problem by limiting license requests on het jobs to the
leader only. Previously, it was only documented that this was a limitation
(see commit a956143202).

Bug 10004