If a cluster is adding itself to the dbd don't send an update back.
So if a slurmctld comes up and finds it is not already in the database
we add it and we were trying to send those associations back.  The
problem with that is the rpc_mgr isn't up yet on the slurmctld yet, so
that connection is deigned.

So here we send a flag into the mysql from the dbd telling it this so
we avoid it all together.

Bug 9140