openapi - slurmdb/v0.0.39/jobs UID/GID resolution
Given the parameters groups/users, it can traverse the list of values
and determine for each ID if it's already a GID/UID, or it's a
group/user name.

IDs are left untouched, whereas names are tried to be substituted
with their IDs. This is done by checking against the system
groups/names, and if it's not a valid one the RESTD query will fail.

Bug 13241