cgroup/v1 - use a better value for STOP_OOM macro
It is okay to set random values on a define constant, but depending on the
type of the variable which it gets assigned to, it can get different values.
Specially in this case, the STOP_OOM value was passed through a pipe and
read on the other side so the other side must know exactly the type it has
to use. It is better to inherently avoid a mistake by just setting a small
value in the define.

Bug 15016