Testsuite - Improve {backup,restore}_config_file avoiding warnings
In the python testsuite, warnings would be seen if an operation
occurred that required a configuration file to be backed up where
the file had already been backed up. Since this can happen under
normal circumstances, the warning can be misleading. I have lowered
the messages to TRACE level so that these messages will not appear
unless you really are looking for the lowest level debug details.

Bug 14923