slurmrestd - use HTTP_STATUS_CODE_ERROR_UNPROCESSABLE_CONTENT
Return HTTP_STATUS_CODE_ERROR_UNPROCESSABLE_CONTENT on
ESLURM_REST_INVALID_QUERY

slurmrestd was able to parse the request but found the content to be
invalid. This is a more accurate status code to give than
HTTP_STATUS_CODE_ERROR_BAD_REQUEST which suggests the request was
malformed or somehow unparsable.