Show correct features in user command after job_array_split
Regression from commit 9b2c3d68cb. features_use is used in
packing/unpacking the job. Since it was set to NULL for the new job
record NULL after the job array was split, the user command displayed
NULL for the features until the next time the job was evaluated by the
scheduler.

Bug 15092