Move data_init call to controller initializations.
The data_init call could happen if someone turned on power saving
capabilities during a reconfig. This was crashing the controller.

Since the data_g_serialize() call can fail gracefully (which disables
support for the SLURM_RESUME_FILE support), we can stop explicitly
requiring the json-c library be available in the slurmctld.

Bug 14925