Improve error handling for send_libs
Change how this error is handled. Failing to get the library files no
longer returns an error (even though there is an error message). In
addition, the BCAST_FLAG_SEND_LIBS flag is unset so that slurmd does not
create a "<filename>_libs" directory.

Clarify the error message when --send-libs=yes or BcastParameters=send_libs
fails to get shared library files: the file is still sent but the shared
library files are not. This is particularly useful when
BcastParameters=send_libs but sbcast is used to send files that don't have
shared libraries (non-executable and non-.so files). In that case, the user
must specify --send-libs=no to avoid the error message.

Bug 15132