scrontab - don't set argc/argv off of the entry command.
Previously, clients displaying Command information for scrontab jobs
would display the name of the entry command and arguments separated by
spaces or tabs, since argv[0] is xstrdup'd from the entry->command, and
the latter set as the result of parsing the whole line as-is.

We want the Command to be displayed as the command name for salloc/srun
and sbatch <script> (without --wrap), otherwise display null. scrontab
conceptually resembles sbatch --wrap in that a command/script isn't
really provided in the command line, but a job specification is edited
and wrapped up in an scrontab entry and job descriptor. Also, argc/argv
aren't actually used for scrontab jobs.

Bug 14331