Allow node suspend/resume timers to be set to lower values
e.g.
PartitionName=p1 SuspendTime=100
PartitionName=p2 SuspendTime=200

If a node is in p1, then it's SuspendTime will be 100.
If the node is added to p2, it will have SuspendTime 200.
If the node transitions back to p1, the SuspendTime would stay at 200.

This is because power_save_set_timeouts() keeps the higher limits. So we
need to clear all values before setting the values again.

Bug 15243