Add flag ignore_prefer_validation to not validate --prefer on a job.
If set, and a job requests --prefer any features in the request that would
create an invalid request with the current system will not generate an error.
This is helpful for dynamic systems where nodes with features come and go.
Please note using this option will not protect you from typos.

Bug 14821