Fix slurmctlds in a federation not recovering a lost connection
If a network issue caused slurmctlds in a federation to lose a
connection, _persist_callback_fini was called and the control_host and
control_port were cleared. This meant that the slurmctlds could never
reestablish that connection until a restart or a federation update
(in which case the control_host and control_port were reestablished).

This commit makes sure that if a connection is lost temporarily (not due
to a shutdown of the daemon) that slurmctld maintains knowledge of how
to recontact that slurmctld and will continue trying to recontact that
slurmctld. This is a partial revert of commit 11d28eabcb from bug 4341,
which also breaks bug 4341 - user commands will hang if a slurmctld
in the federation shuts down.

Bug 11230