Fix slurmctlds in a federation being notified of disconnected ctlds
The previous commit is a partial revert of commit 11d28eabcb from bug
4341, which also breaks bug 4341 - user commands will hang if a
slurmctld in the federation shuts down.

This commit addresses that problem:
With 11d28eabcb, the controller would clear the ControlHost and
ControlPort when a sibling cluster would go away. With the previous
commit this doesn't happen. This commit solves it by having the slurmdbd
update the controllers in the federation when the controller disconnects
from the slurmdbd. Client commands won't contact a cluster with a blank
ControlHost.

Bug 11230