Remove TOCTOU when reading cgroup file interfaces
Getting the size and after reading the file creates a TOCTOU. A file in the
cgroup subsystem is rarely longer than 4KB, and a read over one cgroup
interface should be atomic. Change the functions which read with this unsafe
way into reading by chunks of 4KB till EOF or error.

Bug 15049