Fix accrue time underflow errors after slurmctld reconfig or restart
After a slurmctld reconfig or restart, _restore_job_accounting() is
called and assoc usage info is cleared. Then for each job, if
ACCOUNTING_ENFORCE_LIMITS is set, the usage is restored. For pending
jobs following a reconfig or restart, this accrue usage was not being
properly restored and so future calculations could lead to an underflow
condition.

See also bug 14006.

Bug 15156