Change gpu_index variable name to gpu_minor
The lines in /proc/driver/nvidia-caps/mig-minor are predetermined based on the
GPU's minor number (/dev/nvidia[minor_number]), not the index "i" used in
_get_system_gpu_list_nvml() which correlates to the GPU index as shown by
nvidia-smi. When using multiple GPU's, the minor number is not guaranteed to
match the GPU index, and may only match by chance.

This commit only includes a variable name change.

Bug 15259