Add InfluxDBTimeout parameter to acct_gather.conf
This parameter sets the CURL_TIMEOUT which instructs the CURL API library to
abort the transfer if too much time has elapsed on a sample sent. This helps
on unresponsive servers, specially when the job is terminating and trying to
send the last sample. In that case the job remained in CG until the
UnkillableStepTimeout drained the node.

Bug 15285