sview - fix create partition default toggle problem
In sview, when creating a partition, fix problem when toggling back to default
values for those options that have one listed. Previously, when setting
a new (non-default value) for those options but then when toggling back
(in the pull down menu) to the default one, the default one would not stick.

Bug 15250