switch/hpe_slingshot - instant on OAUTH authentication
PE-44114 - instant on: implement OAUTH authentication

Add code to slurmd to support OAUTH authentication;
if jlope_auth=OAUTH, the instant on init routine will
look at the client-id/client-secret/endpoint files in
the jlope_authdir, and will use them to contact the OAUTH
token service and get a token, which is put into an
Authorization: Bearer header (and cached).
That header is then used to contact the jackalope daemon
for instant on info.

Tested on sawmill (with no OAUTH service) and slice with OAUTH.

Bug 14604