Add tres_per_node to resource_allocation_response_msg
Previously when a pending srun job's GRES was updated with scontrol, the
updated GRES would not be set in the step that is created. This meant that
the job's step would have the incorrect GRES. Now, when the controller
allocates resources for a job, it will also send updated GRES info that is
in job_ptr->tres_per_node.

Bug 13899