Don't overwrite cont_id in cgroup/v2
proctrack_p_create is in charge of setting the container id of the step
and in cgroup/v2 we were setting it too when doing a step_create,
which is wrong.

In Cray, if using task/cgroup it is possible the container id is overwritten
by cgroups and it could race with proctrack/cray_aries which got the cont_id
from the libjob.so from Cray.

Bug 15373