Change srun to wait indefinitely for its allocation
This only changes behavior for srun when it is run outside of an
allocation.

Previously, the default max sleep time was set at (SuspendTimeout +
ResumeTimeout) * 5.  As of this patch, it has been changed to a fixed
300 seconds. Also change backoff behavior to increment by 0.5 seconds
instead of 0.1, but keep the initial 0.1 second sleep for allocations
that are ready very quickly.

Bug 15596