Testsuite - Improve tests avoiding impossible deadline/timelimit combos
Tests 1.104 and 15.32 had jobs with a deadline that was impossible to meet
given the jobs time limit.
In newer versions of Slurm (73815d50127de7) such jobs are rejected upfront,
regardless of the partition time limit.
As there is no reason for those tests to do so, we now use reasonable time
limits.

Bug 15625