Prevent cloud nodes from being rebooting for features while powered up
Normally, when a node becomes idle it is avaiable to reboot into other
active features. With cloud nodes, we want the node to keep the active
features until it is powered off.

_build_node_list() overrides the node_set_ptr's flags to be
NODE_SET_POWER_DN if the node is powered down. Therefore it won't hit
this NODE_SET_REBOOT logic which filters out the non-idle and cloud
nodes.

Bug 14529