Add ability to specify features for specific nodes in helpers.conf
Similar to gres.conf.

Slurmd's will load in all global features and features specific to the
node name. The slurmctld will load in all features.

Defining a feature in the helpers.conf makes it a changeable feature
even if it isn't attached to a specific node in the helpers.conf. For
example, a feature 'fa' can be defined in helpers.conf for node1 and in
the slurm.conf 'fa' can be defined on node2. 'fa' will act as a defined
changeable feature even though 'fa' isn't reported by the helpers
plugin.

e.g.
NodeName=n[1-10] Feature=f1 Helper=/path

Bug 14529