backfill - check bf_max_job_test before going to next_task of array
Because the handling of job array is simplified we're not performing all
checks. This means that we may skip test for number of jobs tested
against bf_max_job_test. If we succeed to schedule job array element, we
need to check if the limit isn't hit before going to next task.

Bug 15685