Validate reported active features
Active features are set on the node when the node is allocated and the
nodes are told to reboot into the active features. The slurmd should
then report the active features that the node has.

Previously it was possible for a node to come up with different active
features than the jobs expected to get by rebooting the node.

If the node doesn't report features then the controller is just managing
the active features.

Bug 14529