Fix regression when updating unsorted nodelists with addrs/hostnames
Introduced in 8a282e6ca4, which causes the NodeName=<nodelist> to be
both sorted and uniq, after taking care of ALL keyword, nodelists, and
nodesets. If NodeAddr and/or NodeHostname were specified to be updated
for the given nodelist, the intended comma-separated 1:1 matching
relation could not match depending on the sorting/uniq changing the
original <nodelist> specified in the node update request.

Bug 15731