Add logic to convert a job features expression to matching OR
This logic converts a job features expression to one where all features
inside of parentheses are AND'd and all the parentheses expressions are
OR'd. This assumes that OR is actually a 'matching OR' where all nodes in
the job must have the same set of features rather than a boolean OR where
different nodes in the job can have a different set of features.

Bug 12286